Update patch set 4
Patch Set 4:

This is (sort of) expected, because there are existing bugs related to option parsing.

Current trunk abe does not support the gdbserver= option, it is erroneously accepted as an alias for gdb=. Thus your command line has the effect of setting unmatched versions of gdb and gdbserver. In this situation, trunk abe will either fail to build gdbserver, or silently build a mismatched version of gdbserver depending on the contents of the snapshots directory.

The scope of the patch is to ensure that the assumptions made by abe (in particular the checkout() function, but secondarily configure_build()) are not violated. This error message correctly indicates that something has already gone wrong before checkout() was called.

Patch-set: 4