Update patch set 2
Patch Set 2:

the abe-tests fail because of this error: "configure: error: You need to install these packages:  gcc-multilib g++-multilib". This is a different bug as Ubuntu xenial doesn't need these packages anymore. ABE configure should not make this an error for Xenial. This of course has nothing to do with git submodules.

Patch-set: 2
Label: Code-Review=0