Update patch set 1
Patch Set 1: Code-Review+1

As discussed in the team lead's call, we don't want to use ${host} because this would mean that for the Linux builds there might be two different manifests generated: on named based on x86_64 as a host and one named on i686 as a host.  The only difference would be the name of the file, not the contents.

Patch-set: 1