Update patch set 1
Patch Set 1: Code-Review+2

Charles,
The Foundation model is installed automatically on new builders as part of https://ci.linaro.org/view/tcwg-ci/job/tcwg-update-snapshots_ref/ .  This job is kept up-to-date with the list of machines.

Christophe,
I don't know why we wanted to have it in /linaro.

Patch-set: 1
Label: Code-Review=+2