Update patch set 1
Patch Set 1:

(1 comment)

This change has merit, but using stage1_flags instead of default_configure_flags seems wrong.

Also, we need to make it in a way that only affect gcc-6-based toolchains, and not gcc-5 to avoid adding new features to a maintenance toolchain.  This means that the change needs to go to "config/gcc6/gcc.conf".

Patch-set: 1
Label: Code-Review=0