Update patch set 5
Patch Set 5:

"very likely" is what seems. The entire discussion hasn't been decided yet. As it stands, this is what's going to happen:

End Aug: All proposals should be finalised. Sub-modules is up, mono-repo is getting ready, there might be more.
Early Sep: We'll put a survey to choose which way is forwards. The survey will take ~ 1 months to complete, with opinions from all over the community.
Early Oct: We'll close the survey and start analysing the data. We want to be as thorough as possible, until "just before" the US LLVM Meeting, crunching as much data as we can.
Early Nov: During the US LLVM Meeting, we'll have a session to present the results and hopefully take the decision.

The decision might be anything from staying as it is, to changing radically to any of the proposals in the survey. There's no way to tell at this stage.

Patch-set: 5
Reviewer: Renato Golin <1000088@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0