Update patch set 1
Patch Set 1:

(3 comments)

I support this idea, but think the implementation is not ideal because it is difficult to reliably discover whether the network is available.

I think the real problem is the use of "git ls-remote" to do URL validation. This means that:
1) the network is required for all builds, even when sources are already downloaded
2) it is not possible to build when the upstream git server is down, even if the sources are already downloaded.

Rather than try to make the network detection more reliable, it would be better to stop abe from using the network when it is not really necessary.

If you move the "git ls-remote" check to line 203, does that solve the offline use case (without needing the network detection)?

Patch-set: 1
Reviewer: Charles Baylis <1000108@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0