Update patch set 1
Patch Set 1:

I think the only reason we needed these package for x86_64 was due to an issue fixed in https://review.linaro.org/#/c/13717/ .  In that issue the multilib was enabled for x86_64 native build, which required system to have multilib support packages installed.  Once the above patch is merged (and multilib disabled for x86_64 native), there is no need to install or check for gcc-multilib and g++-multilib.

Patch-set: 1