scripts/MakeRelease.job: Match the naming scheme used in the Jenkins job.
We want to upload releases of different gcc branches to different
directories: 5-2016.08, 6-2016.08 for instance.

2016.08 is provided in Jenkins by the release manager, but the
gcc major version is extracted from the tarball names. Use the
same logic as in the Jenkins job.

There is still an unclear code condition, which might be cleaned
if useless. In doubt, I added a comment before it.

Change-Id: I9da981ea74c457184b4ef15459598a201a4238bc