Warn when dryrun'd commands fail.
Since commit 04adeb37 (Report unhandled errors), unhandled errors
cause the build to fail, but it is not easy to identify the failing
command. Display a warning when dryrun detects that a command failed.
This is reported as a warning because the real error reporting/handling
comes later, and the build may continue if the error handling is
missing.

Change-Id: Ia4d5919ae228bd3a562d545767ab09e8afe7824f