Update patch set 2
Patch Set 2:

The glaring hole in the logic is that gcc_src is not a required input parameter for MakeRelease.job, and thus can't be relied upon for parsing the GCC version information.  I've resubmitted a patch which parses the GCC version from the generated manifest file name.  I'm going to test this with a make release build.

Patch-set: 2
Label: Code-Review=0