Update patch set 5
Patch Set 5:

It looks like abe already allows tag names to be used in place of branch names, although https://review.linaro.org/#/c/14045/ is required to suppress an error which is displayed when doing so.

Most other changes in this pull request to stamps, manifests, etc do not seem to be necessary.

Patch-set: 5