Update patch set 1
Patch Set 1:

The reason I asked about libiberty and such is that you mention standard configure practice, so I would expect libiberty to be configured in a similar way to zlib, and thus have host multilibs.

I've looked at the Makefiles and configures, but so far I couldn't understand why they behave differently. In particular, it's not clear if it's a bug or intended purpose.

It looks like the easiest way is to modify abe such that it builds zlib, and uses that as "system" lib.

Patch-set: 1