jenkins.sh: Split build and test into 2 abe.sh invocations
The purpose is two-fold.  Firstly, this patch is a step towards
separating build and test stages, which we need for testing of
binary releases.

Immediatelly, however, this patch makes --testcontainerfile
jenkins.sh option possible.  By splitting test from the build
we give time for $testcontainerfile to appear once the test container
is started.

Change-Id: Ie85fffa8f46498f08fc2eb2fcdd5733586826a1e