jenkins.sh: Pass --testcontainer option to abe.sh
... when --testcontainerfile option is given to jenkins.sh.

It can take up to 20 minutes to start a testing container when the
testing board is under heavy load, so it is best to parallelize it
with the build of the toolchain.  If starting a container exceeds
time required to build the toolchain -- then we need to investigate
and fix the bottlenecks on the tester side.

Change-Id: Iba061228a9aec90e8d7f9e71162526685a48d8b4