Update patch set 3
Patch Set 3:

Charles,

This patch doesn't change how cross-toolchains are built, it merely moves --disable-multilib setting for them to a different place.

The patch is no-op only for the native toolchains -- which we do not release, so no need for changes to extraconfigdirs for gcc-4_9 and gcc-5.

Patch-set: 3
Label: Code-Review=0