Update patch set 2
Patch Set 2:

Thanks for the hint Maxim. The comparison I posted is misleading in that it compares a build with abe-stable and abe-master+this patch.

abe-master enabled ifunc, and also upgraded to binutils-2.27 (stable still points to 2.26), so this explains the differences.

But indeed, I saw gcc/g++ ifunc tests failing, but I didn't have the time to look at why yet.

Patch-set: 2
Label: Code-Review=0