Update patch set 1
Patch Set 1: Code-Review+1

It looks sane.
But I think this is not a problem in practice, since we call this script with ${srcs} pointing to the tag or top of release branch (which does not change during a release build).

But for manual tests, it can be a problem.

Patch-set: 1