Fix sysroot path when importing manifest.
When abe is run to build from a manifest in a directory which contains
the name of the selected target, abe fails to add the target name
to the sysroot path. This causes abe-tests failures, and could also
cause invalid sysroots to be created if multiple targets are built
in the same snapshots directory.

This is caused by a conditional in import_manifest which does not
append the target name to ${sysroots} if it already contains the
target name. This was presumably done to avoid appending the target
name twice if "--target xxx --manifest" was added to the command line
but this is now forbidden by abe since commit ID:
3d8110557edaa8c115d7f037f785b39e1ab999b6 Reject use of --target with
--manifest.

Change-Id: I98d76c476296b8e3f10266d1c7fef0b63f4f3c78