Refactor parts of build_all() into separate functions.
build_all() does a lot more than building. This is a partial refactor
to separate the make check and tarball generation into separate
functions. Later cleanup will move the calls to these functions into
a separate top-level control function.

The poor error handling of do_tarsrc and do_tarbin is retained. It
will be cleaned up in a later patch.

Change-Id: Ic52843862c4d03209db311bb984ae84316507a0b