Update patch set 1
Patch Set 1:

Don't we need to disable the built-in zlib before we commit this patch?

Playing devil's advocate...

Isn't the real problem that gcc's configure passes the --enable-multilib to zlib, and could/should we fix that instead with a gcc patch?

Does relying on the system zlib introduce a compatibility regression for users who are using distros other than Trusty (older or newer)?

Patch-set: 1