Update patch set 1
Patch Set 1:

I was under the impression that some of the changes in the qemu snapshot had not made it into the 2.7 release and that we would need to wait for 2.8 before going back to mainline qemu.

I don't remember this clearly, so I could be wrong, but I think it's worth checking.

The abe-tests failure is a network timeout and unrelated to the patch.

Patch-set: 1