Update patch set 1
Patch Set 1:

My proposal is to move QEMU binaries away from public view to dev-01.tcwglab.
QEMU is has implementations of crypto, which is export-controlled.  This is one of the reasons why Linaro's virtualization team does not provide binaries.  We could do the whole crypto-providing-NSA-notification dance (it's not difficult), but that is a separate effort.

For now, I'll update this patch to pull binaries from dev-01.tcwglab.

Patch-set: 1
Label: Code-Review=0