Update patch set 1
Patch Set 1:

gcc configure doesn't look at anything past the -linux.  The hf at the end is just an end user convention to indicate that the toolchain was configured --with-float=hard.  This is not something that gcc configure enforces.

I tried building an arm-linux-gnueabihf toolchain with your patch, and I got a toolchain configured --with-float=soft which is the default for all arm-linux-* toolchains.

This patch looks wrong to me.

Patch-set: 1
Reviewer: Jim Wilson <1000132@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0