Improve hello_world test.
The hello_world test is taken out of the build step, and put
into it's own build step. This means that the sysroot is fully
installed and it is not necessary to use the --sysroot option
when performing the test.

In addition, the test is enhanced to work for newlib-based
arm and aarch64 targets.

Change-Id: I35adb580d411a10caa47f6052ded3afb16e90ad8