lib/component.sh: Support new git-branch output format.
After checking out a revision, 'git branch' version 1.9.1 from trusty
used to report:
* (detached from 833c70b)

As of version 2.7.4 (trusty), the same command generates:
* (HEAD detached at 833c70b)

Since we used to check the presence of detached in the 2nd
field, we get '(HEAD' instead of the expected '(detached'
which confuses collect_data/component_init.

This patch fixes this by adding '^(HEAD' in the expected
strings.

Change-Id: I856f200ebee72fca38b0074bb49549c291b6140d