Update patch set 1
Patch Set 1: Code-Review+2

I agree, and I have a plan to remove scripts/test-schroot.sh and most of lib/schroot.sh.

For a variety of reasons we want to add custom compilation flags when running testsuites (e.g., test non-default -mcpu/-mfpu settings).  When this is implemented, we can add "-Wl,-dynamic-linker=<path>/ld.so -Wl,-rpath=<path>/libc" and just rsync sysroot to <path> on the target.  With support for custom paths to sysroot we can remove test-schroot.sh and schroot.sh as that's the only task they are currentlty doing.

Patch-set: 1
Label: Code-Review=+2