Add CPPFLAGS to gdbserver configure line.
Since gdbserver is built before the INSTALL_SYSROOT step, it
is necessary to use --sysroot when invoking the compiler. We
already set CFLAGS and CXXFLAGS, but also need to set CPPFLAGS
so that the configure script can correctly find header files for
testing platform features.

Testing to see if this fixes build failures on arch.

Change-Id: I0fddc6ea8be0ac553a84c16470aca6b66f955c50