Clean up version parsing from command line and .conf files.
The parsing of the ${latest} variable which is set in the .conf
files, and the parsing of the command line component version
strings is confusing, (re-)using poorly named variables and setting
unused variables. Clean this up so it is easier to read.

Also, only prepend the component name to the version for tarballs.

Change-Id: Ie2ce0a96ac6dbb9f2851045662abe017c3652e85