Remove broken "depends" variable from conf files.
The "depends" mechanism appears to have been intended to
store separate dependency lists for each component, but in fact
is stored in a single global variable. Thus, the value of
${depends} was always set to the value in the last component
*.conf file to be sourced (gdb or gdbserver). Since both gdb and
gdbserver have depends="" in their conf files, the feature is
useless.

However, Abe separately parses the depends line in
infrastructure.conf, so that line is retained.

Change-Id: I45413bba69b23e26bd04df4e8a41afca59a16d91