Update patch set 2
Patch Set 2:

OK, but then I need to update this patch for consistency: we'd use a linaro branch for binutils, and not for glibc.
We'd have to formalize how we maintain our branches (currently our binutils-2.27 branch at least is behind fsf-2.27, lacking a few patches for non-arm targets).

I'm also wondering about newlib.... do we stick to a version, to avoid blindly following trunk? If yes, which version?

Patch-set: 2