Update patch set 2
Patch Set 2:

Sorry, I'm still not sure to understand :-)
For the time being, no backport is needed in glibc nor binutils, so why create a local branch? In practice we tend to forget to update it (merge with upstream). Or, for homogeneity, create a local branch for every component, and setup a formal update process (like for gcc, with monthly merges with upstream). IMO, having different policies for different components is confusing and error-prone.

Patch-set: 2
Label: Code-Review=0