Update patch set 1
Patch Set 1:

manifest test failures are expected.

wrt. keeping it enabled for our current branches, we can do that, but since objc is only built in our validation for native builds and not part of our releases I don't think it's worth it to continue testing it.

Patch-set: 1
Label: Code-Review=0