Update patch set 1
Patch Set 1:

As mentioned in https://review.linaro.org/#/c/18124/, configure does not complain for host and target to be different when bootstrap is enabled and the build fails at stage-2. Perhaps it might be a good idea to check that host and target are same for a bootstrap build (normal or specialized configs like bootstrap-lto).

Patch-set: 1
Reviewer: Prathamesh Kulkarni <1000119@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0