Update patch set 1
Patch Set 1:

> "Build Failed
 > 
 > https://ci.linaro.org/jenkins/job/tcwg-abe-test-gerrit-cambridge/993/
 > : Well ... it worked on my machine"

This seems to be failed at the command:

"17:33:24 find: `/home/tcwg-buildslave/workspace/tcwg-abe-test-gerrit-cambridge_2/label/tcwg-x86_64-build/test/native_build/_build/builds/x86_64-unknown-linux-gnu/x86_64-unknown-linux-gnu/glibc.git~linaro~2.21~master/posix/globtest-dir/noread': Permission denied"

Which is ran after the glibc make check from:

17:17:32 RUN: make check RUNTESTFLAGS=" --xml=glibc.xml "  CFLAGS_FOR_BUILD=-"-pipe" CXXFLAGS_FOR_BUILD="-pipe" -j 32 -w -i -k -C /home/tcwg-buildslave/workspace/tcwg-abe-test-gerrit-cambridge_2/label/tcwg-x86_64-build/test/native_build/_build/builds/x86_64-unknown-linux-gnu/x86_64-unknown-linux-gnu/glibc.git~linaro~2.21~master 2>&1 | tee /home/tcwg-buildslave/workspace/tcwg-abe-test-gerrit-cambridge_2/label/tcwg-x86_64-build/test/native_build/_build/builds/x86_64-unknown-linux-gnu/x86_64-unknown-linux-gnu/glibc.git~linaro~2.21~master/check-glibc.log
17:33:23 NOTE: Test run completed after 54 minutes
17:33:23 NOTE: Performing build step INSTALL_SYSROOT
17:33:23 NOTE: Performing build step HELLO_WORLD
17:33:23 RUN: (hello_world)
17:33:23 NOTE: Hello World test succeeded
17:33:23 NOTE: Complete build process took 54 minutes
17:33:23 ++ jenkins.sh: 399: find /home/tcwg-buildslave/workspace/tcwg-abe-test-gerrit-cambridge_2/label/tcwg-x86_64-build/test/native_build/_build/builds/ -name destdir -prune -o -name '*manifest.txt' -print
17:33:24 find: `/home/tcwg-buildslave/workspace/tcwg-abe-test-gerrit-cambridge_2/label/tcwg-x86_64-build/test/native_build/_build/builds/x86_64-unknown-linux-gnu/x86_64-unknown-linux-gnu/glibc.git~linaro~2.21~master/posix/globtest-dir/noread': Permission denied

Which puzzles me because this directory is create and removed by a glibc own testcase (globtest.sh) using a trap in the shell script.  So it could be case where either the shell script than actually ran it (glibc issues the default one /bin/sh in this case) failed to act on the trap or failed to remove the directory. I will check on how to filter this out or see if this is a glibc issue.

Patch-set: 1