Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > looks OK to me, except for the config change.
 > 
 > I think we need to fix the Jenkins wipe-workspace problem before we
 > apply this though.

 > (1 comment)
 > 
 > looks OK to me, except for the config change.
 > 
 > I think we need to fix the Jenkins wipe-workspace problem before we
 > apply this though.

I do not have a strong opinion here, although the config is generation an invalid libc.so shared object that render it segfault at the start (due invalid fields in init_array segment).

I will push a patch on glibc to force it cleanup the tests files for a segfaults.

Patch-set: 4