Update patch set 3
Patch Set 3:

Hi,
Sorry for late reply. The issue here is that gdbserver expects CXX to be a cross compiler for ARM and by overriding CXX="g++ -O2 -fsanitize=undefined" we are passing it the host compiler, which results in the build failure. I think this is more of a user error in that, if one expects to build gdbserver then one should not override CXX and set it to host compiler anyway
(since gdbserver is an ARM binary). Also the use-case for overriding CC/CXX is just for building gcc itself with ubsan only (stage-1).
I was wondering whether we should instead add an option say --stage1_ubsan to abe that would build gcc with ubsan (and rest of the components built as they are) ?

Thanks,
Prathamesh

Patch-set: 3