Update patch set 3
Patch Set 3:

would it work to configure GCC with:

   .../configure ... CC='gcc -fsanitize=undefined' CXX='g++ -fsanitize=undefined'

If it does, it could be done with an extraconf gcc.conf which adds CC=... CXX=... to the default_configure_flags. This could be in the extraconfs repo, and wouldn't need Abe changes.

Patch-set: 3
Reviewer: Charles Baylis <1000108@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=0