Remove git ls-remote check from checkout() and retrieve()
The purpose of the git ls-remote is to check the URL for validity.
Since stderr is directed to /dev/null, a failure of this check is
hard to debug. The check is unnecessary, as the URL is checked later
by git clone, so it can be removed.

In order to placate the testsuite, a rudimentary URL syntax check
is included.

Change-Id: Ia03bf176b78c72ecbba73594e48d87dcc42b5260