Fix bug #3040: CC1 and cc1plus cannot convert UTF-8
When building a toolchain for a mingwin host, use a
pre-built libiconv.

Because a bug in gcc-6, we also have to use --disable-nls instead of
--enable-nls. The problem was fixed by r226712, by the part that
updates intl/configure. However, that patch seems too invasive to
backport it to our linaro-gcc-6-branch, and also seems to have other
dependencies. --enable-nls works with gcc-7 and trunk, but it does not
seem worth enabling/disabling it depending on which branch is being
built.

We also update the manifest version to 1.5 to account for this new
component.

Change-Id: I6b7ab0c8b62c452815570d9a2de1ba702ac31dad