Update patch set 2
Patch Set 2:

Charles, you are correct, this approach is essentially the same as test-schroot.sh's in respect that ssh and bash will start using new sysroot when it is installed (I removed that part from commit message, but didn't highlight removal).

I considered the approach you are suggesting, but didn't get to trying it.  Let me give it a go.

Patch-set: 2