Update patch set 1
Patch Set 1: Code-Review+1

The abe-tests-checkout failures are "expected": we rebuild the toolchain with the same command-line as without this patch, and compare the manifest. Obviously, they differ:
-gdb_branch=gdb-8.0-branch
+gdb_branch=gdb-8.1-branch
-gdbserver_branch=gdb-8.0-branch
+gdbserver_branch=gdb-8.1-branch

We have this difference for gcc5 and gcc6 toolchains.
For gcc-4.9, the abe config forces gdb-7.1

For "some reason", both gcc-4.9 and gcc-7 lack gdbserver info in the manifest: I don't know if it's actually built or not.

Patch is OK, but the gdbserver issue is worth a look....

Patch-set: 1
Reviewer: Christophe Lyon <1000072@f33910f19b7abb192b83adbd000000bf>
Label: Code-Review=+1