Update patch set 1
Patch Set 1:

An example usage would be:
--send-results-cmd bsd-mailx --send-results-to christophe.lyon@linaro.org
provided that email is properly configured on the machine.
I used bsd-mailx because it has the same user-interface as the default Mail command used by the contrib/ script.

The intended use is in our buildfarm containers, for instance from the trunk monitoring jobs.

Patch-set: 1