Fix usage of abe.sh --checkout
Contrary to what the usage says, --checkout does not allow to set a
branch or revision for a component, it only allows to select the
component to checkout. However, the optional parameters at the end which
allow to specify a package that differs from the default version allows
to specify a branch or a revision using the same syntax as described in
the --checkout option. This commit fixes the usage accordingly.

Change-Id: I7c0c996e0ec4824d1379767017df3e664cf09372