globals.sh: Rename $gcc into $hostgcc
The rest of ABE assumes gcc[] is an associative array, and defining a
scalar variable named $gcc make __[gs]et_component_GENERIC fail to
detect that the 'gcc' component does not exist (for instance when
building a subset of a toolchain, not including gcc).

This patch just changes the name of this temporary variable.

Change-Id: Ifec54f1d7a762ad7fb9bc0c0a67014e88ae5ded4