Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> (1 comment)
> 
> Shouldn't you also update gcc.conf (which uses $bootstrap, which you patch removes) ?
> What about lib/configure.sh?
re: gcc.conf
I added bootstrap to lib/globals.sh as a "bool" to check if bootstrap is enabled or disabled. By default it is set to "no".
Should we instead remove all uses of $bootstrap and test build_config != "" when we need to check if bootstrap is enabled or is the current version OK ?

Patch-set: 5
Reviewer: Gerrit User 1000119 <1000119@f33910f19b7abb192b83adbd000000bf>