Build a C++ compiler during stage1 when using glibc
Since release 2.29, glibc requires a C++ compiler, so build one when
building a toolchain that uses glibc.

This fixes a bug where we would do this only when building gcc master,
while this in fact depends on the glibc version we build.

We cannot patch the gcc stage1 build flags as part of glibc.conf, because
.conf files are sourced in isolation, forbidding overwriting flags of
other components (as a safety/sanity feature).

Change-Id: Ie9e363354c3362a160d15d13f49f93556cc0d790