Update patch set 2
Patch Set 2:

Just to make it clear, glibc does not really require a C++ compiler. What glibc requires is a CXX set correctly for cross-compilation.  The way we usually cross-compiling a complete toolchain for glibc testing is better described at glibc script build-many-glibcs.py [1], but in a short:

1. There is no need to enable C++ compiler on stage1, so I think the current setup is correct (only enable c++ on stage2).

2. For glibc configure you will need to set a CXX based on the build stage1 compiler, even if is an invalid one (for instance, aarch64-linux-gnu-g++ pointing to an invalid path or not accessible in $PATH). 

The issue that glibc is throwing now is it uses autogen AC_PROG_CXX and it tries for g++ in the system path and if it builds/links a simple test it will end up using it. For cross-compilation this is incorrect, so by passing an invalid CXX compiler glibc will disable CXX usage.

[1] https://sourceware.org/git/?p=glibc.git;a=blob;f=scripts/build-many-glibcs.py;h=c5821df25e0fa7e70d0e37b385b7d207b710cc14;hb=HEAD

Patch-set: 2
Reviewer: Gerrit User 1000232 <1000232@f33910f19b7abb192b83adbd000000bf>