Update patch set 2
Patch Set 2:

The tests fail because the manifest is now different since the stage1 flags change with this patch.

Regarding Adhemerval's comments, since we want to run glibc tests, we could consider that a C++ compiler is required in our case. However, IIUC, we currently only run glibc tests for native builds (in which case I suspect we use the system C++ compiler, which may or may not be what we want).
If we always define CXX to a non-existing compiler, will that always disable all the tests, or only some of them?

Patch-set: 2
Reviewer: Gerrit User 1000072 <1000072@f33910f19b7abb192b83adbd000000bf>