lib/configure.sh: configure_build() Define CXX when configuring glibc
Since version 2.29, glibc can make use of a C++ compiler when running
its tests, but can incorrectly use the host's C++ compiler when
cross-compiling, leading to a build failure. Defining CXX to
${target}-c++ fixes this, and is OK even if that compiler doesn't
exist (in which case glibc doesn't try to use it)

When building a cross-toolchain, we'll point to a non-existing
compiler, unless we are building gcc master branch where we build a
C++ compiler during stage1.

When building a native toolchain, we'll usually use the system's C++
compiler as it is in the PATH at configure time.

Change-Id: I415c8cb77377c106d551bc4e2c68f6cb7fb2d457