Add ability to choose --with-multilib-list profile for arm newlib
NOTE: This patch does not necessarily need to be merged as it adds a
feature we may not need for regular use of abe. It is being sent for
review to preserve if we need in the future.

To build the equivalent of the GNU RM Embedded toolchain we need to use
the rmprofile multilibs. To build these we need to use
--with-multilib-list=rmprofile. We also want to change the default
configure options for newlib to match those in the GNU RM Embedded
toolchain, for example disable threading.

We implement with a new command line option:
[--set {multilib}={aprofile|rmprofile}]
This is only valid when newlib is the clibrary and an arm target is
selected. This changes the --with-multilib-list to the appropriate profile
and selects the appropriate configure options for newlib. We also need to
alter the hello world sanity check to use --cortex-m3 --specs=rdimon.specs

Change-Id: I6adcb85fab890450b59c3cbab2d31f93729fcb00