Update patch set 2
Patch Set 2:

In the previous case it was easy to avoid the manifest change.  In this case we would need to fix up ABE to implement this.
If (and I stress, "if") we want to avoid changing the manifest, one solution could be to add "glibc_needs_cxx=true" flag to config/master/glibc.conf, and then check for it in config/master/gcc.conf .
In practice, the manifest change will be harmless, so we may just accept it.
FWIW, either way is fine with me.

Patch-set: 2
Reviewer: Gerrit User 1000092 <1000092@f33910f19b7abb192b83adbd000000bf>