Make abe to write results of test_summary to result.txt.
If --check gcc is passed, then use contrib/test_summary script to
obtain test results summary and write it to result.txt, which can
then be used as input to compare_tests script.

The patch could be refactored to avoid calling test_summary twice,
but I didn't attempt that in this patch.

Change-Id: Ifee6cf2be907bcfdc0b2a8aa84be0947bce87866