Update patch set 3
Patch Set 3:

(1 comment)

I've fixed the test failure, and have updated the default in the help.

The test failure was specific to AArch64 as in that case we don't have a profile so the config got missed. I've updated that in config/newlib. There may be a better way of doing that to avoid auto being there. Perhaps changing to (pseudocode) if ${multilib_list} == rmprofile then ... else ...

Patch-set: 3
Reviewer: Gerrit User 1000216 <1000216@f33910f19b7abb192b83adbd000000bf>