Update patch set 1
Patch Set 1:

The idea with this patch is just to workaround the problem with gdb-8.3*. Do you think there can be an upstream patch to fix that branch?
Anyway, it would be good to have our builds succeed again until that happens, if it happens.

Patch-set: 1
Reviewer: Gerrit User 1000072 <1000072@f33910f19b7abb192b83adbd000000bf>