Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The idea with this patch is just to workaround the problem with gdb-8.3*. Do you think there can be an upstream patch to fix that branch?
> Anyway, it would be good to have our builds succeed again until that happens, if it happens.

I think we can make a case for including it in gdb-8.3 once there is a fix. The problem is the code is a bit gnarly and there are other offenders that trigger errors.

I think the fix makes sense for getting our builds going again. But then we'd revert it once a fix is available. Does that sound ok?

Patch-set: 1
Reviewer: Gerrit User 1000373 <1000373@f33910f19b7abb192b83adbd000000bf>