Always "touch" auto-generated files before build
Before this patch we update timestamps on files only when we
create a new source directory.  With increased usage of ABE
from round-robin scripts, which handle source checkouts themselves
(due to need to run "git bisect" and such) we need to update ABE
to play better in environments where it doesn't handle sources.

Change-Id: Ic01b43080a6ab183b954fd8ce3b1746d7d0a6803