Move $sysroots under $prefix
This is the first patch in a series to rework handling of sysroot
in ABE.  At the moment sysroot is made by copying bits and pieces
as the build progresses from various places: GCC libraries, Glibc
install location, etc.  This is more complicated than needed.

The primary issue stems from a decision to build target sysroot
outside of $prefix.  GCC toolchain build machinery has an interesting
property that if a directory (headers, libraries, sysroot) is specified
as a path inside GCC's $prefix, then GCC will use relative paths to
access it.  I speculate that this little-known fact was not taken into
account and sysroot was configured to be on-the-side.

This patch sets $sysroots to be inside $prefix, and cleans up handling
of both $prefix and $sysroots variables along the way.

Change-Id: I3d4126cbb169310a030c1f3e9e87e6b7ed699eea