config/gcc.conf: Don't set default multilib setting if the user has overridden them
Without this patch, if the user uses --set gcc_override_configure=--disable-multilib
GCC for arm-eabi is effectively configured with:
--enable-multilib --with-multilib-list=aprofile --disable-multilib

If the user also overrides --with-cpu, GCC's consistency checks fail
because --with-multilib-list=aprofile is still taken into account and
considered conflicting with --with-cpu (--disable-multilib does not
remove multiib-list).

This patch avoids setting the detault multilib flags if the user has
overridden then, so for instance the above example would result in
configuring with --disable-multilib --with-cpu=XXX, which would pass
GCC consistency checks.

Change-Id: I67c29de3165599c2dcda4a64aaf135d228f67b5e