config/gdb.conf: Set configure flag '--with-auto-load-safe-path'
This allows GDB to load libthread_db.so.1 from our sysroot.

Unfortunately the original value for auto-load-safe-path contains '$'
characters, which are very hard to preserve while the configure flags move
through the innards of Abe. To solve this problem, use '@@' in their place
and change configure_build() to substitute '@@' for '$' at the point where
the configure flags are used.

Change-Id: Ic8563f121d9781d9b6e949e477722be96252ed33