Add --rerun-failed-tests option
The purpose is to weed out flaky tests and allow the CI to flag
regressions.

The sum and log files of each run are preserved in files with .<n>
extension, where <n> is the number of the attempted run, starting with 0.

A few command line options are also added:

--gcc-compare-results option is needed so that its compare_dg_tests.pl can
be used to merge the sum files into a synthetic sum file where a test is
marked as passed if it passes in any of the testsuite reruns.

--ci-project and --ci-config are used to find the relevant flaky tests list
for the current build.

Co-developed-by: Maxim Kuvyrkov <maxim.kuvyrkov@linaro.org>
Change-Id: I97c5bf61fba48cfae4446ab127ddb13ac58ba141