lib/make.sh (make_all): Use '--specs=rdimon.specs' for aarch64*-elf GCC
As discussed in https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66203
aarch64*-none-elf toolchains using newlib need the
--specs=rdimon.specs option otherwise link fails because _exit
etc... cannot be resolve.

This has a side-effect in libstdc++ configure, which detects the
compiler cannot build executables, and thus defines gcc_no_link=yes,
which in turn means that the detection of getentropy() becomes a
compile-only test.

Unfortunately, newlib declares getentropy() via unistd.h, but provides
no implementattion, which means that libstdc++ incorrectly guesses
that getentropy() is available.

In turn, this means that linking with libstdc++ pulls a dependency
only getentropy() which results in link failure.  An earlier version
of this problem was discussed in
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=103166, and a proper fix
would be to implemented a getentropy() stub in newlib.

In the mean time, we instead add --specs=rdimon.specs to
LDFLAGS_FOR_TARGET for aarch64*elf, next to
-Wl,--fix-cortex-a53-843419.

Maybe we should add a cleaner support to override LDFLAGS_FOR_TARGET,
and add these flags in newlib.conf, but this would be a more intrusive
patch.

Change-Id: I1695492376977191e00ae0162535e80012fbc742